requirements.txt
Django
Django-crispy-forms
python-slugify
pip install crispy-bootstrap4